pyproject.toml
[project]
name = "template-aqa-project"
version = "0.1.0"
description = "Template project for aqa, used pytest, playwright, requests, allure"
authors = [
    { name = "Anton Bogomolov", email = "ashteriel@ya.ru" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[tool.poetry]
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-xdist = "^3.6.1"
pytest-playwright = "^0.6.2"
allure-pytest = "^2.13.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"

[tool.poetry.group.linter.dependencies]
ruff = "^0.8.6"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"


[tool.pytest.ini_options]
testpaths = [
    "src.api.tests",
    "src.ui.tests",
    "src.e2e.tests",
]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
norecursedirs = [".git", ".venv"]
python_files = "test_*.py"
python_functions = "test_*"
merkers = [
    "regression: Регрессионные тесты"
]
addopts = [
    "-v",
    "-s"
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 80
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.black]
line-length = 79
target-version = ["py313"]

[tool.isort]
profile = "black"
py_version = "3.13"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
